+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## Unreleased
+- Fixed data path (correct path: `~/.local/share/satnogs-tabulation-helper/`)
+- Fixed marker input (rounded to nearest integer pixel value).
+- Added STRF rffit hint text (by using `--show-hint`).
 
 ## 0.2